adds migration for tweets

+++ db/migrate/20140812222654_add_uid_to_tweets.rb
@@ -1,5 +1,5 @@
 class AddUidToTweets < ActiveRecord::Migration
   def change
-    add_column :tweets, :uid, :integer
+    add_column :tweets, :uid, :bigint
   end
 end
